backend/pyproject.toml
[tool.poetry]
name = "test-api"
version = "0.1.0"
description = "Test Api by Duc Anh Le"
authors = ["Duc Anh Le"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
prisma = "^0.15.0"
pydantic = "1.10.6"
starlette = "^0.46.1"
celery = "^5.4.0"
redis = "^5.2.1"
python-dotenv = "^1.0.1"
ujson = "^5.10.0"
fastapi-event = "^0.1.3"
click = "^8.1.8"
greenlet = "^3.1.1"
httpx = "^0.28.1"
PyYAML = "^6.0.2"
email-validator = "^2.2.0"
python-jose = "^3.4.0"
cryptography = "^44.0.2"
passlib = "^1.7.4"
bcrypt = "^4.3.0"
Faker = "^37.0.2"
pytest-asyncio = "^0.25.3"
exceptiongroup = "^1.2.2"
uvloop = "^0.21.0"
watchfiles = "^1.0.4"
anyio = "^4.9.0"
flower = "^2.0.1"
loguru = "^0.7.3"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pylint = "^3.3.5"
pytest = "^8.3.5"
behave = "^1.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
